backend/requirements.txt
Flask
flask-cors
openai
google-generativeai
sentence-transformers
faiss-cpu
pdfplumber
docx2txt
unstructured
reportlab
